[2025-06-15, sungwoong-onboarding] 응답지 제출 기능 및 테스트

+++ project/sungwoong-onboarding/src/main/java/com/multi/sungwoongonboarding/options/infrastructure/OptionsJpaRepository.java
@@ -2,5 +2,5 @@ package com.multi.sungwoongonboarding.options.infrastructure;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OptionsJpaRepository extends JpaRepository<OptionsJpaEntity, Integer> {
+public interface OptionsJpaRepository extends JpaRepository<OptionsJpaEntity, Long> {
 }
